requirements.txt
# ==============================================================================
# CVFlix Backend - Requirements
# Python 3.11+
# Instalación: bash install.sh (automático)
# ==============================================================================

# ==================== BUILD TOOLS ====================
setuptools>=69.0.0,<70.0.0
wheel>=0.42.0,<1.0.0

# ==================== BASE ====================
numpy>=1.25.2,<2.0.0

# ==================== SYSTEM MONITORING ====================
psutil>=5.9.0,<6.0.0

# ==================== FACE RECOGNITION ====================
# ⚠️ IMPORTANTE: dlib NO está en este archivo
# El instalador lo maneja automáticamente:
#   - conda install -c conda-forge dlib (wheels precompilados)
# cmake>=3.18.0
# dlib>=19.24.0

face-recognition>=1.3.0,<2.0.0
face-recognition-models>=0.3.0

# ==================== SERVER & API ====================
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
websockets>=12.0,<13.0
sse-starlette>=1.8.0,<2.0.0
httpx>=0.25.1,<0.28.0
requests>=2.31.0,<3.0.0

# ==================== VIDEO PROCESSING ====================
opencv-python-headless>=4.8.0,<5.0.0
Pillow>=10.1.0,<11.0.0

# ==================== MACHINE LEARNING ====================
# ⚠️ TensorFlow NO está aquí - el instalador lo maneja por sistema
# Razón: pip no siempre detecta bien los markers en conda
# El script install.sh instala la versión correcta automáticamente

scikit-learn>=1.3.0,<1.5.0
scipy>=1.11.0,<1.14.0

# ==================== VISUALIZATION ====================
matplotlib>=3.8.0,<4.0.0

# ==================== EXTERNAL API ====================
webcolors>=1.13.0,<2.0.0

# ==================== VALIDATION & CONFIG ====================
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# ==================== SECURITY ====================
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
cryptography>=41.0.0,<43.0.0